+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Actualize `nodejs` version in CI
 * Migrate to `rails-7` and it's configuration defaults
 * Check `dependabot` at 8:00 Moscow time daily
+* Fix `rubocop-rails` issues after upgrade to `rubocop-rails v2.14.0`
 
 ### Fixes